参数调整

+++ dao-mybatis/src/main/java/com/chenbo/daomybatis/controller/UserController.java
@@ -17,7 +17,7 @@ public class UserController {
     private UserService userService;
 
     @RequestMapping("{id}")
-    public String getUser(@PathVariable int id) {
+    public String getUser(@PathVariable double id) {
         return userService.selectById(id).toString();
     }
 }
